requirements.txt
algoliasearch==1.15.3
algoliasearch-django==1.5.5
amqp==2.2.2
apipkg==1.4
attrs==17.4.0
billiard==3.5.0.3
boto3==1.5.27
botocore==1.8.41
certifi==2017.11.5
chardet==3.0.4
cookies==2.2.1
coverage==4.4.2
Django==2.0.2
django-cors-headers==2.1.0
django-guardian==1.4.9
django-model-utils==3.1.1
django-storages==1.6.5
djangorestframework==3.7.7
docutils==0.14
ephem==3.7.6.0
execnet==1.5.0
factory-boy==2.8.1
Faker==0.8.8
flake8==3.5.0
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
hyperlink==17.3.1
idna==2.6
inflection==0.3.1
iso8601==0.1.12
jmespath==0.9.3
kombu==4.1.0
mccabe==0.6.1
msgpack-python==0.5.4
pbr==3.1.1
pep8==1.7.1
pluggy==0.6.0
promise==2.1
psycopg2==2.7.3.2
py==1.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
pytest==3.4.0
pytest-cache==1.0
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-env==0.6.2
pytest-factoryboy==1.3.1
pytest-flake8==0.9.1
pytest-mock==1.6.3
pytest-pep8==1.0.6
python-dateutil==2.6.1
pytz==2017.3
raven==6.5.0
redis==2.10.6
requests==2.18.4
responses==0.8.1
Rx==1.6.0
s3transfer==0.1.12
singledispatch==3.4.0.3
six==1.11.0
text-unidecode==1.1
typing==3.6.2
urllib3==1.22
vine==1.1.4
websocket-client==0.46.0
zope.interface==4.4.3
